完善webdav备份功能

+++ lib/views/settings/webdav.dart
@@ -1,13 +1,17 @@
 import 'dart:io' as io;
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
+import 'package:haka_comic/config/app_config.dart';
 import 'package:haka_comic/database/history_helper.dart';
 import 'package:haka_comic/database/images_helper.dart';
 import 'package:haka_comic/database/read_record_helper.dart';
+import 'package:haka_comic/utils/common.dart';
+import 'package:haka_comic/utils/log.dart';
 import 'package:haka_comic/widgets/button.dart';
 import 'package:haka_comic/widgets/toast.dart';
 import 'package:path_provider/path_provider.dart';
 import 'package:webdav_client/webdav_client.dart';
+import 'package:path/path.dart' as p;
 
 enum ActionType { upload, download }
 
@@ -19,11 +23,21 @@ class WebDAV extends StatefulWidget {
 }
 
 class _WebDAVState extends State<WebDAV> {
+  final appConf = AppConf();
   final _urlController = TextEditingController();
   final _usernameController = TextEditingController();
   final _passwordController = TextEditingController();
   ActionType _actionType = ActionType.upload;
   bool _loading = false;
+
+  @override
+  void initState() {
+    super.initState();
+
+    _urlController.text = appConf.webdavUrl;
+    _usernameController.text = appConf.webdavUser;
+    _passwordController.text = appConf.webdavPassword;
+  }
 
   @override
   void dispose() {
@@ -40,22 +54,29 @@ class _WebDAVState extends State<WebDAV> {
     final url = _urlController.text.trim();
     final username = _usernameController.text.trim();
     final password = _passwordController.text.trim();
+
+    if (url.isEmpty || username.isEmpty || password.isEmpty) {
+      Toast.show(message: '请填写完整信息');
+      return;
+    }
+
     var client = newClient(
       url,
       user: username,
       password: password,
       debug: kDebugMode,
-    );
+    )..setHeaders({"Content-Type": "application/octet-stream"});
+
     setState(() {
       _loading = true;
     });
+
     try {
       final list = await client.readDir('/');
       if (_actionType == ActionType.upload) {
         if (list.any((item) => item.name == 'HaKa Comic')) {
           await client.remove('/HaKa Comic/');
         }
-        // await client.mkdir(baseDir);
         final files = await Future.wait([
           ImagesHelper.backup(),
           HistoryHelper().backup(),
@@ -63,14 +84,16 @@ class _WebDAVState extends State<WebDAV> {
         ]);
         await Future.wait(
           files.map((file) {
-            final fileName = file.path.split('/').last;
+            final fileName = p.basename(file.path);
             return client.writeFromFile(file.path, '$baseDir/$fileName');
           }),
         );
-        Toast.show(message: '上传成功');
         for (var file in files) {
-          await file.delete();
+          if (await file.exists()) {
+            await file.delete();
+          }
         }
+        Toast.show(message: '上传成功');
       } else if (_actionType == ActionType.download) {
         if (!list.any((item) => item.name == 'HaKa Comic')) {
           Toast.show(message: '远程目录不存在');
@@ -99,9 +122,11 @@ class _WebDAVState extends State<WebDAV> {
         ]);
         Toast.show(message: '下载成功');
       }
+
+      save();
     } catch (e) {
-      print('WebDAV error: $e');
-      Toast.show(message: '连接失败');
+      Log.error('WebDAV error', e);
+      showSnackBar('发生错误: ${e.toString()}');
     } finally {
       setState(() {
         _loading = false;
@@ -109,72 +134,83 @@ class _WebDAVState extends State<WebDAV> {
     }
   }
 
+  void save() {
+    appConf.webdavUrl = _urlController.text.trim();
+    appConf.webdavUser = _usernameController.text.trim();
+    appConf.webdavPassword = _passwordController.text.trim();
+  }
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(title: const Text('WebDAV')),
-      body: ListView(
-        padding: const EdgeInsets.all(12),
-        children: [
-          TextField(
-            decoration: const InputDecoration(
-              labelText: 'URL',
-              hintText: '例如：https://example.com/webdav',
-              border: OutlineInputBorder(),
-            ),
-            controller: _urlController,
-          ),
-          const SizedBox(height: 12),
-          TextField(
-            decoration: const InputDecoration(
-              labelText: '用户名',
-              border: OutlineInputBorder(),
-            ),
-            controller: _usernameController,
-          ),
-          const SizedBox(height: 12),
-          TextField(
-            decoration: const InputDecoration(
-              labelText: '密码',
-              border: OutlineInputBorder(),
-            ),
-            controller: _passwordController,
-          ),
-          const SizedBox(height: 12),
-          Row(
-            mainAxisSize: MainAxisSize.min,
+      body: Center(
+        child: ConstrainedBox(
+          constraints: const BoxConstraints(maxWidth: 400),
+          child: ListView(
+            padding: const EdgeInsets.all(12),
             children: [
-              const Text('操作'),
-              Radio(
-                value: _actionType,
-                groupValue: ActionType.upload,
-                onChanged: (value) {
-                  setState(() {
-                    _actionType = ActionType.upload;
-                  });
-                },
+              TextField(
+                decoration: const InputDecoration(
+                  labelText: 'URL',
+                  hintText: '例如：https://example.com/webdav',
+                  border: OutlineInputBorder(),
+                ),
+                controller: _urlController,
               ),
-              const Text('上传'),
-              const SizedBox(width: 12),
-              Radio(
-                value: _actionType,
-                groupValue: ActionType.download,
-                onChanged: (value) {
-                  setState(() {
-                    _actionType = ActionType.download;
-                  });
-                },
+              const SizedBox(height: 12),
+              TextField(
+                decoration: const InputDecoration(
+                  labelText: '用户名',
+                  border: OutlineInputBorder(),
+                ),
+                controller: _usernameController,
               ),
-              const Text('下载'),
+              const SizedBox(height: 12),
+              TextField(
+                decoration: const InputDecoration(
+                  labelText: '密码',
+                  border: OutlineInputBorder(),
+                ),
+                controller: _passwordController,
+              ),
+              const SizedBox(height: 12),
+              Row(
+                mainAxisSize: MainAxisSize.min,
+                children: [
+                  const Text('操作'),
+                  Radio(
+                    value: _actionType,
+                    groupValue: ActionType.upload,
+                    onChanged: (value) {
+                      setState(() {
+                        _actionType = ActionType.upload;
+                      });
+                    },
+                  ),
+                  const Text('上传'),
+                  const SizedBox(width: 12),
+                  Radio(
+                    value: _actionType,
+                    groupValue: ActionType.download,
+                    onChanged: (value) {
+                      setState(() {
+                        _actionType = ActionType.download;
+                      });
+                    },
+                  ),
+                  const Text('下载'),
+                ],
+              ),
+              const SizedBox(height: 12),
+              Button.filled(
+                onPressed: connectWebDAV,
+                isLoading: _loading,
+                child: const Text('继续'),
+              ),
             ],
           ),
-          const SizedBox(height: 12),
-          Button.filled(
-            onPressed: connectWebDAV,
-            isLoading: _loading,
-            child: const Text('继续'),
-          ),
-        ],
+        ),
       ),
     );
   }
